requirements.txt
scapy>=2.5.0
PyYAML
pydantic
fastapi
uvicorn[standard]
